# YJIT: Allow inlining ISEQ calls with a block