--- a/benchmark/loop_times_megamorphic.yml
+++ b/benchmark/loop_times_megamorphic.yml
@@ -1,0 +1,7 @@
+prelude: |
+  eval(<<~EOS)
+    def loop_times_megamorphic
+      #{"1.times {|i|};" * 1000}
+    end
+  EOS
+benchmark: loop_times_megamorphic